requirements.txt
torch==2.0.1
transformers==4.36.0
sentence-transformers==2.3.1
huggingface-hub==0.20.0
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
gradio==4.12.0
accelerate==0.25.0
bitsandbytes==0.41.3
pydantic==2.5.0
pandas==2.1.4
